feat(attachment-picker): reduce cognitive complexity
